renaming _latlonzoom to _geo
