+++ sanitiser/reverse.js
@@ -2,7 +2,7 @@
 var logger = require('../src/logger'),
     _sanitize = require('../sanitiser/_sanitize'),
     sanitiser = {
-      latlonzoom: require('../sanitiser/_latlonzoom'),
+      latlonzoom: require('../sanitiser/_geo'),
       size: function( req ) {
         var size = require('../sanitiser/_size');
         return size(req, 1);
